When duplicating DBs, handle the case of :memory: paths

+++ sotodlib/core/metadata/manifest.py
@@ -382,7 +382,11 @@ class ManifestDb:
         of writing a Db to disk to call copy(map_file=...) and then
         simply discard the returned object.
         """
-        if map_file is not None and os.path.exists(map_file):
+        if (
+            map_file is not None and 
+            map_file != ":memory:" and
+            os.path.exists(map_file)
+        ):
             if overwrite:
                 os.remove(map_file)
             else:
